Move malicious behavior tests to a submodule
