+++ src/protocols/misbehavior_tests.rs
@@ -1,0 +1,4 @@
+mod aux_gen;
+mod interactive_signing;
+mod key_init;
+mod key_refresh;
