+++ src/protocols/misbehavior_tests/key_init.rs
@@ -8,7 +8,7 @@ use manul::{
 };
 use rand_core::{CryptoRngCore, OsRng};
 
-use super::key_init::{KeyInit, KeyInitAssociatedData, Round2NormalBroadcast, Round3, Round3NormalBroadcast};
+use super::super::key_init::{KeyInit, KeyInitAssociatedData, Round2NormalBroadcast, Round3, Round3NormalBroadcast};
 use crate::{
     curve::Scalar,
     dev::TestParams,
